pyproject.toml
[project]
name = "coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astor>=0.8.1",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.1",
]
